refactor(RankType): 출력과 관련된 기능 이동
- RankType에서 출력과 관련된 기능을 OutputView로 이동하여 책임 분리
